feat(social): merge the Nextdoor setup and settings into one form
