--- plugins/newspack-plugin/src/wizards/newspack/views/settings/social/nextdoor/settings/style.scss
@@ -1,29 +1,0 @@
-/**
- * Nextdoor Settings Component Styles
- */
-
-@use "~@wordpress/base-styles/colors" as wp-colors;
-@use "~@wordpress/base-styles/variables" as wp-vars;
-
-.nextdoor-settings {
-	&__subheading {
-		font-size: wp-vars.$font-size-medium;
-		font-weight: 600;
-		line-height: wp-vars.$font-line-height-medium;
-		margin: 0;
-	}
-
-	&__intro {
-		margin: 0;
-	}
-
-	&__status-value {
-		&--success {
-			color: wp-colors.$alert-green;
-		}
-
-		&--error {
-			color: wp-colors.$alert-red;
-		}
-	}
-}
